+++ sem5/aod/lista3/lib/Graph.hpp
@@ -6,19 +6,19 @@ using namespace std;
 class Graph {
 public:
 	int n; // liczba wierzchołków
-	int m; // liczba krawedzi
-	vector<vector<pair<int,int>>> adj; // lista sąsiedztwa u: (v, cost)
-	int minCost, maxCost;
+	long long m; // liczba krawedzi
+	vector<vector<pair<int,long long>>> adj; // lista sąsiedztwa u: (v, cost)
+	long long minCost, maxCost;
 
-	Graph(int n, vector<pair<pair<int,int>,int>> E); // lista krawędzi ((u,v), cost)
-	Graph(int n, vector<vector<pair<int,int>>> adj);
+	Graph(int n, vector<pair<pair<int,int>,long long>> E); // lista krawędzi ((u,v), cost)
+	Graph(int n, vector<vector<pair<int,long long>>> adj);
 
 	void DijkstraFull(int s);
-	int DijkstraTwo(int s, int t);
+	long long DijkstraTwo(int s, int t);
 
 	void DialFull(int s);
-	int DialTwo(int s, int t);
+	long long DialTwo(int s, int t);
 
 	void RadixHeapFull(int s);
-	int RadixHeapTwo(int s, int t);
+	long long RadixHeapTwo(int s, int t);
 };
