+++ sem5/aod/lista3/lib/Writer.hpp
@@ -2,5 +2,5 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-void writeP2p(string inFile, string outFile, string algorithm, int n, int m, int minCost, int maxCost, vector<pair<pair<int,int>,int>> results);
+void writeP2p(string inFile, string outFile, string algorithm, int n, int m, int minCost, int maxCost, vector<pair<pair<int,int>,long long>> results);
 void writeSS(string inFile, string outFile, string algorithm, int n, int m, int minCost, int maxCost, double avg);
